Need to handle seconds past midnight.

+++ GoldenEmbedParser/src/org/shampoo/goldenembed/parser/SortBySeconds.java
@@ -1,10 +1,8 @@
 package org.shampoo.goldenembed.parser;
 
-public class SortBySeconds implements java.util.Comparator<GoldenCheetah>
-{
-    public int compare(GoldenCheetah aGC, GoldenCheetah anotherGC) 
-    {
-        int secDiff = aGC.getSecs() - anotherGC.getSecs();
+public class SortBySeconds implements java.util.Comparator<GoldenCheetah> {
+    public int compare(GoldenCheetah aGC, GoldenCheetah anotherGC) {
+        int secDiff = (int) (aGC.getSecs() - anotherGC.getSecs());
         return secDiff;
-    } 
+    }
 }
